+++ kubekarma/controlleroperator/httpserver.py
@@ -10,10 +10,8 @@ from fastapi.responses import JSONResponse
 
 from pydantic import BaseModel
 
-from kubekarma.shared.genericcrd import TestCaseResultItem
-from kubekarma.controlleroperator import get_results_publisher
+
 import logging
-from typing import List
 
 
 import uvicorn
@@ -46,17 +44,6 @@ async def validation_exception_handler(request: Request, exc: RequestValidationE
 @app.get("/healthz")
 def healthz():
     return {"status": "ok"}
-
-
-@app.post("/api/v1/execution-tasks/{execution_id}")
-def post_results(execution_id: str, items: List[TestResultsModel]):
-    print("Received results for execution task:", execution_id)
-    # Convert the items to TestResults
-    get_results_publisher().notify_new_results(
-        execution_id,
-        results=[TestCaseResultItem(**item.model_dump()) for item in items]
-    )
-    return {"item_id": execution_id}
 
 
 class ThreadedUvicorn:
