+++ tox.ini
@@ -20,6 +20,7 @@ deps =
     -r{toxinidir}/requirements.controller.txt
     -r{toxinidir}/requirements.worker.txt
     -r{toxinidir}/requirements.test.txt
+    -r{toxinidir}/requirements.dev.txt
 passenv =
     HOME
     PIP_CACHE_DIR
@@ -50,3 +51,7 @@ deps =
     -r{toxinidir}/requirements.test.txt
 commands =
     flake8 {toxinidir}/kubekarma --format='%(path)s:%(row)d,%(col)d %(code)s:%(text)s:https://lintlyci.github.io/Flake8Rules/rules/%(code)s.html' --count
+
+[testenv:lint-imports]
+commands =
+    lint-imports
